requirements.txt
fastapi
uvicorn
python-multipart
pillow
pytesseract
jinja2
python-dotenv
google-generativeai
